POL-96 친구 기능 리팩토링

+++ src/main/java/com/polychat/polychatbe/friend/command/application/dto/FriendUserDTO.java
@@ -9,8 +9,8 @@ import lombok.NoArgsConstructor;
 public class FriendUserDTO {
 
     @NotBlank(message = "친구 정보가 유효한지 확인해주세요.")
-    private final int user1;
+    private final long user1;
 
     @NotBlank(message = "친구 정보가 유효한지 확인해주세요.")
-    private final int user2;
+    private final long user2;
 }
